support config sleep time, the default is 10ms

+++ src/main/java/com/esgyn/kafkaCDC/server/utils/Utils.java
@@ -157,7 +157,7 @@ public class Utils {
 	return name;
     }
 
-    public static void waitMillisecond(int milliseconds) {
+    public static void waitMillisecond(long milliseconds) {
 	try {
 	    Thread.sleep(milliseconds);
 	} catch (Exception e) {
